Refactored the structure code
Structures are now object oriented, and are no longer hardcoded in.
Rendering of structures is done using pointsprites, which kinda works.
Added trees back in, using the same noise function as before. Need to
figure out how to get texture coordinates to work with the point sprites,
possible through a fragment shader.